Delegate getting of stations in viewmodels to common object
